+++ stream-investment/investment-core/src/main/java/com/backbase/stream/investment/InvestmentAssetsTask.java
@@ -2,6 +2,7 @@ package com.backbase.stream.investment;
 
 import com.backbase.investment.api.service.v1.model.AssetCategory;
 import com.backbase.investment.api.service.v1.model.AssetCategoryType;
+import com.backbase.investment.api.service.v1.model.GroupResult;
 import com.backbase.investment.api.service.v1.model.Market;
 import com.backbase.investment.api.service.v1.model.MarketSpecialDay;
 import com.backbase.stream.worker.model.StreamTask;
@@ -36,11 +37,18 @@ public class InvestmentAssetsTask extends StreamTask {
     public void setAssetCategoryTypes(List<AssetCategoryType> assetCategoryTypes) {
         data.setAssetCategoryTypes(assetCategoryTypes);
     }
+
     public void setAssetCategories(List<AssetCategory> assetCategories) {
         data.setAssetCategories(assetCategories);
     }
+
     public void setAssets(List<Asset> assets) {
         data.setAssets(assets);
+    }
+
+    public InvestmentAssetsTask setPriceTasks(List<GroupResult> tasks) {
+        data.setPriceAsyncTasks(tasks);
+        return this;
     }
 
 }