+++ stream-investment/investment-core/src/main/java/com/backbase/stream/investment/saga/InvestmentSaga.java
@@ -3,6 +3,7 @@ package com.backbase.stream.investment.saga;
 import com.backbase.stream.configuration.InvestmentIngestionConfigurationProperties;
 import com.backbase.stream.investment.InvestmentData;
 import com.backbase.stream.investment.InvestmentTask;
+import com.backbase.stream.investment.service.AsyncTaskService;
 import com.backbase.stream.investment.service.InvestmentClientService;
 import com.backbase.stream.investment.service.InvestmentModelPortfolioService;
 import com.backbase.stream.investment.service.InvestmentPortfolioAllocationService;
@@ -64,6 +65,7 @@ public class InvestmentSaga implements StreamTaskExecutor<InvestmentTask> {
     private final InvestmentPortfolioService investmentPortfolioService;
     private final InvestmentPortfolioAllocationService investmentPortfolioAllocationService;
     private final InvestmentModelPortfolioService investmentModelPortfolioService;
+    private final AsyncTaskService asyncTaskService;
     private final InvestmentIngestionConfigurationProperties coreConfigurationProperties;
 
     @Override
@@ -125,13 +127,15 @@ public class InvestmentSaga implements StreamTaskExecutor<InvestmentTask> {
 
     private Mono<InvestmentTask> upsertPortfoliosAllocations(InvestmentTask investmentTask) {
         InvestmentData data = investmentTask.getData();
-        return Flux.fromIterable(Objects.requireNonNullElse(data.getPortfolios(), List.of()))
-            .flatMap(
-                p -> investmentPortfolioAllocationService.generateAllocations(p,
-                    data.getPortfolioProducts(),
-                    investmentTask.getData().getInvestmentAssetData()))
-            .collectList()
-            .map(o -> investmentTask);
+        return asyncTaskService.checkPriceAsyncTasksFinished(data.getPriceAsyncTasks())
+            .then(Flux.fromIterable(Objects.requireNonNullElse(data.getPortfolios(), List.of()))
+                .flatMap(
+                    p -> investmentPortfolioAllocationService.generateAllocations(p,
+                        data.getPortfolioProducts(),
+                        investmentTask.getData().getInvestmentAssetData()))
+                .collectList()
+                .map(o -> investmentTask)
+            );
     }
 
     /**