+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [8.9.0](https://github.com/Backbase/stream-services/compare/8.8.0...8.9.0)
+### Changed
+- Investment service
+  - add support for asset categories
+  - improve allocations setup
+  - add deposit and orders
+  - price ingest async tasks complete check 
+
 ## [8.8.0](https://github.com/Backbase/stream-services/compare/8.7.0...8.8.0)
 ### Changed
 - Investment service